Create (Up|Down)Loader package
