+++ Modules/(Up|Down)Loader/Executers/DownloadExecuter/Background/BackgroundDownloadExecuter.swift
@@ -1,4 +1,5 @@
 import Foundation
+import Jetworking
 
 /**
  * # Summary
