+++ Modules/(Up|Down)Loader/Executers/UploadExecuter/Background/BackgroundUploadExecuter.swift
@@ -1,4 +1,5 @@
 import Foundation
+import Jetworking
 
 final class BackgroundUploadExecuter: NSObject, UploadExecuter {
     var delegate: UploadExecuterDelegate?
