+++ Modules/(Up|Down)Loader/Executers/UploadExecuter/Default/DefaultUploadExecuter.swift
@@ -1,4 +1,5 @@
 import Foundation
+import Jetworking
 
 final class DefaultUploadExecuter: NSObject, UploadExecuter {
     var delegate: UploadExecuterDelegate?
