+++ Modules/(Up|Down)Loader/Executers/UploadExecuter/UploadExecuterType.swift
