+++ Package.swift
@@ -10,20 +10,25 @@ let package = Package(
         .iOS(.v10)
     ],
     products: [
-        // Products define the executables and libraries produced by a package, and make them visible to other packages.
         .library(
             name: "Jetworking",
             targets: ["Jetworking"]),
+        .library(
+            name: "(Up|Down)Loader",
+            targets: ["(Up|Down)Loader"])
     ],
     targets: [
-        // Targets are the basic building blocks of a package. A target can define a module or a test suite.
-        // Targets can depend on other targets in this package, and on products in packages which this package depends on.
         .target(
             name: "Jetworking"
         ),
+        .target(
+            name: "(Up|Down)Loader",
+            dependencies: ["Jetworking"],
+            path: "Modules/(Up|Down)Loader"
+        ),
         .testTarget(
             name: "JetworkingTests",
-            dependencies: ["Jetworking"],
+            dependencies: ["Jetworking", "(Up|Down)Loader"],
             resources: [
                 .copy("Resources/avatar.png")
             ]
