+++ Sources/Jetworking/Client/HTTP/HTTPMethod.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-enum HTTPMethod: String {
+public enum HTTPMethod: String {
     case GET = "GET"
     case POST = "POST"
     case PUT = "PUT"
